+++ picoquic/util.c
@@ -229,9 +229,9 @@ void picoquic_parse_packet_header_cnxid_lengths(uint8_t l_byte, uint8_t *dest_le
     *srce_len = (h2 == 0) ? 0 : h2 + 3;
 }
 
-uint32_t picoquic_format_connection_id(uint8_t* bytes, size_t bytes_max, picoquic_connection_id_t cnx_id)
+uint8_t picoquic_format_connection_id(uint8_t* bytes, size_t bytes_max, picoquic_connection_id_t cnx_id)
 {
-    uint32_t copied = cnx_id.id_len;
+    uint8_t copied = cnx_id.id_len;
     if (copied > bytes_max || copied == 0) {
         copied = 0;
     } else {
@@ -249,7 +249,7 @@ int picoquic_is_connection_id_length_valid(uint8_t len) {
     return ret;
 }
 
-uint32_t picoquic_parse_connection_id(const uint8_t * bytes, uint8_t len, picoquic_connection_id_t * cnx_id)
+uint8_t picoquic_parse_connection_id(const uint8_t * bytes, uint8_t len, picoquic_connection_id_t * cnx_id)
 {
     if (picoquic_is_connection_id_length_valid(len)) {
         cnx_id->id_len = len;
